hw12_calendar замечания
