+++ hw12_13_14_15_calendar/cmd/calendar/main.go
@@ -11,13 +11,9 @@ import (
 	"github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/internal/app"
 	"github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/internal/config"
 	"github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/internal/logger"
-	httpserver "github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/internal/server"
 	internalhttp "github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/internal/server/http"
-	"github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/internal/storage"
-	"github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/internal/storage/database"
-	memorystorage "github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/internal/storage/memory"
-	sqlstorage "github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/internal/storage/sql"
 	"github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/pkg/shortcuts"
+	"github.com/Lanworm/OTUS_GO/hw12_13_14_15_calendar/pkg/storageConf"
 )
 
 var configFile string
@@ -42,23 +38,7 @@ func main() {
 	logg, err := logger.New(config.Logger.Level, os.Stdout)
 	shortcuts.FatalIfErr(err)
 
-	var eventStorage storage.IStorage
-	if config.Storage.InDatabase() {
-		db := database.New(&config.Database)
-		err := db.Connect(ctx)
-		shortcuts.FatalIfErr(err)
-
-		defer func() {
-			err := db.Connect(ctx)
-			if err != nil {
-				logg.Error(err)
-			}
-		}()
-
-		eventStorage = sqlstorage.NewEventStorage(db, logg)
-	} else {
-		eventStorage = memorystorage.New()
-	}
+	eventStorage, _ := storageconf.NewStorage(ctx, *config, logg)
 
 	calendar := app.New(logg, eventStorage)
 
@@ -80,16 +60,10 @@ func main() {
 	}()
 
 	logg.Info("calendar is running...")
-	handler := new(httpserver.Handler)
-	registerRoutes(server, handler)
 
 	if err := server.Start(ctx); err != nil {
 		logg.Error("failed to start http server: " + err.Error())
 		cancel()
 		os.Exit(1) //nolint:gocritic
 	}
-}
-
-func registerRoutes(srv *internalhttp.Server, handler *httpserver.Handler) {
-	srv.AddRoute("/hello", handler.Hello)
 }
